Remove old nav
